Added jwt tokens

+++ frontEnd/leaguelizer/src/Pages/CompetitionPage/CustomForm.js
@@ -1,8 +1,9 @@
-import { React, useEffect, useState, useCallback } from "react";
+import { React, useEffect, useState, useCallback, useContext } from "react";
 import { Button, Grid, TextField, Autocomplete, Container } from "@mui/material";
 import URL_BASE from "./constants";
 import CustomTable from "../../Layouts/PageLayout/Table/CustomTable";
 import ToasterError from "../../Layouts/ErrorLayout/ToasterError";
+import authContext from "../../Context/Context";
 
 const initialMatchValue = {
     "club1": {
@@ -27,6 +28,7 @@ const initialMatchValue = {
 }
 
 export default function CustomForm(props) {
+    let {user} = useContext(authContext);
     const [compNameValue, setCompNameValue] = useState(props.value.name);
     const [compTeamValue, setCompTeamValue] = useState(props.value.numberOfTeams);
     const [compDateValue, setCompDateValue] = useState(props.value.foundedDate);
@@ -56,7 +58,7 @@ export default function CustomForm(props) {
     const [ orderDirection, setOrderDirection ] = useState("asc");
     const [ pageNumber, setPageNumber ] = useState(1);
     const [ pageMax, setPageMax ] = useState(1);
-    const [ paginationValue, setPaginationValue ] = useState(12);
+    const [ paginationValue, setPaginationValue ] = useState(localStorage.getItem('paginationValue') ? JSON.parse(localStorage.getItem('paginationValue')) : 12);
 
     function validateCompetition() {
         if (!/^[0-9]+$/.test(compTeamValue) | parseInt(compTeamValue) < 0){
@@ -65,6 +67,18 @@ export default function CustomForm(props) {
         }
         if (!/^[0-9]+$/.test(compPrizeValue) | parseInt(compPrizeValue) < 0){
             ToasterError("There must be a positive prize");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]+$/.test(compNameValue)){
+            ToasterError("Competition Name can only contain numbers and letters");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]+$/.test(compTypeValue)){
+            ToasterError("Competition type can only contain numbers and letters");
+            return false;
+        }
+        if (!/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/.test(compDateValue)){
+            ToasterError("Date needs to have the following format: yyyy-mm-dd");
             return false;
         }
         return true;
@@ -79,17 +93,66 @@ export default function CustomForm(props) {
             ToasterError("The team cannot play with itself!");
             return false;
         }
+        if (!/^[0-9]+$/.test(club1Value) && matchValue.club1.id === ""){
+            ToasterError("Invalid club1 id");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(club2Value) && matchValue.club2.id === ""){
+            ToasterError("Invalid club2 id");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(stadiumValue) && matchValue.stadium.id === ""){
+            ToasterError("Invalid stadium id");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]*$/.test(roundValue)){
+            ToasterError("Invalid Round of play");
+            return false;
+        }
+        if (!/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/.test(dateValue)){
+            ToasterError("Date needs to have the following format: yyyy-mm-dd");
+            return false;
+        }
+        return true;
+    }
+
+    function validateClub() {
+        if (!/^[0-9]+$/.test(clubAnnualBudgetValue) | parseInt(clubAnnualBudgetValue) < 0){
+            ToasterError("Annual Budget must be a positive integer");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(clubStaffValue) | parseInt(clubStaffValue) < 0){
+            ToasterError("Number of staff must be a positive integer");
+            return false;
+        }
+        if (!/^[a-zA-Z0-9 ]+$/.test(clubNameValue)){
+            ToasterError("Club Name can only contain numbers and letters");
+            return false;
+        }
+        if (!/^[0-9]+$/.test(clubStadiumValue)){
+            ToasterError("Club stadium incorrect!");
+            return false;
+        }
+        if (!/^[0-9]{4}-[0-9]{2}-[0-9]{2}$/.test(clubDateValue)){
+            ToasterError("Date needs to have the following format: yyyy-mm-dd");
+            return false;
+        }
         return true;
     }
 
     const getPageMax = useCallback(() => {
         if (props.value.id === undefined)
             return;
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
     
-        fetch(URL_BASE + String(props.value.id) + "/clubs/?pageNumber=0")
+        fetch(URL_BASE + String(props.value.id) + "/clubs/?pageNumber=" + String(paginationValue))
             .then(number => number.json())
             .then(number => setPageMax(number["pageNumber"]));
-    }, [props.value.id]);
+    }, [props.value.id, paginationValue]);
 
     useEffect(() => {
         setCompNameValue(props.value.name)
@@ -105,6 +168,11 @@ export default function CustomForm(props) {
         if (props.value.id === undefined)
             return "";
 
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return "";
+        }
+
         return URL_BASE + String(props.value.id) + "/clubs/?page=" + String(pageNumber) + "&pageNumber=" + String(paginationValue);
     }, [pageNumber, props.value.id, paginationValue])
 
@@ -116,7 +184,12 @@ export default function CustomForm(props) {
     }, [getUrlForMatches])
 
     useEffect(() => {
-        if (specificLeagueVisible && compNameValue !== ""){
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
+        
+        if (specificLeagueVisible && compNameValue !== "" && props.value.id !== undefined){
             fetch(URL_BASE + String(props.value.id))
                 .then(clubs => clubs.json())
                 .then(clubs => setLeagueClubs(clubs["clubs"]));
@@ -201,15 +274,18 @@ export default function CustomForm(props) {
                 "numberOfTeams": compTeamValue,
                 "foundedDate": compDateValue,
                 "prizeMoney": compPrizeValue,
-                "competitionType": compTypeValue
+                "competitionType": compTypeValue,
+                "user":(user) ? user.user_id : null
             })
         };
 
         fetch(URL_BASE, requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.user !== undefined)
+                    ToasterError(message.user[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
                 else
                     props.refresh();
             })
@@ -218,6 +294,11 @@ export default function CustomForm(props) {
     const putButtonHandler = () => {
         if (!validateCompetition())
             return;
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
 
         const requestOptions = {
             method: 'PUT',
@@ -236,14 +317,21 @@ export default function CustomForm(props) {
         fetch(URL, requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.user !== undefined)
+                    ToasterError(message.user[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
                 else
                     props.refresh();
             })
     }
 
     const deleteButtonHandler = () => {
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
+
         const requestOptions = {
             method: 'DELETE'
         };
@@ -251,13 +339,7 @@ export default function CustomForm(props) {
         const URL = URL_BASE + String(props.value.id)
 
         fetch(URL, requestOptions)
-            .then(message => message.json())
-            .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
-                else
-                    props.refresh();
-            })
+            .then(() => {props.refresh();});
     }
 
     const postWithClubsHandler = () => {
@@ -273,43 +355,48 @@ export default function CustomForm(props) {
                 "foundedDate": compDateValue,
                 "prizeMoney": compPrizeValue,
                 "competitionType": compTypeValue,
+                "user":(user) ? user.user_id : null,
                 "clubs": leagueClubs
             })
         };
 
-        fetch(URL_BASE + "leagueClubs", requestOptions)
+        fetch(URL_BASE + "leagueClubs/", requestOptions)
             .then(message => message.json())
             .then((message) => {
                 if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                    ToasterError(message.error);
                 else
                     props.refresh();
             })
     }
 
     const addClubHandler = () => {
+        if (!validateClub()){
+            return;
+        }
+
         let varLeagueCLubList = leagueClubs;
         const newCLub = {
             "name": clubNameValue,
             "annualBudget": clubAnnualBudgetValue,
             "numberOfStadd": clubStaffValue,
             "foundedDate": clubDateValue,
-            "stadium": clubStadiumValue
+            "stadium": clubStadiumValue,
+            "user":(user) ? user.user_id : null
         }
+
         varLeagueCLubList.push(newCLub);
         setLeagueClubs(varLeagueCLubList);
-    }
-
-    const refresh = () => {
-        changeMatchValues(initialMatchValue);
-        fetch(getUrlForMatches())
-            .then(match => match.json())
-            .then(match => setMatchList(match));
     }
 
     const postMatchButtonHandler = () => {
         if (!validateMatch())
             return;
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
 
         const requestOptions = {
             method: 'POST',
@@ -320,23 +407,39 @@ export default function CustomForm(props) {
                 "stadium": (/^[0-9]+$/.test(stadiumValue)) ? stadiumValue : matchValue.stadium.id,
                 "roundOfPlay": roundValue,
                 "score": scoreValue,
-                "date": dateValue
+                "date": dateValue,
+                "user":(user) ? user.user_id : null
             })
         };
 
         fetch(URL_BASE + String(props.value.id) + "/clubs/" , requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.club1 !== undefined)
+                    ToasterError(message.club1[0]);
+                else if (message.club2 !== undefined)
+                    ToasterError(message.club2[0]);
+                else if (message.stadium !== undefined)
+                    ToasterError(message.stadium[0]);
+                else if (message.competition !== undefined)
+                    ToasterError(message.competition[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
+                else if (message.user !== undefined)
+                    ToasterError(message.user);
                 else
-                    refresh();
+                    props.refresh();
             })
     }
 
     const putMatchButtonHandler = () => {
         if (!validateMatch())
             return;
+
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
 
         const requestOptions = {
             method: 'PUT',
@@ -357,14 +460,29 @@ export default function CustomForm(props) {
         fetch(URL, requestOptions)
             .then(message => message.json())
             .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
+                if (message.club1 !== undefined)
+                    ToasterError(message.club1[0]);
+                else if (message.club2 !== undefined)
+                    ToasterError(message.club2[0]);
+                else if (message.stadium !== undefined)
+                    ToasterError(message.stadium[0]);
+                else if (message.competition !== undefined)
+                    ToasterError(message.competition[0]);
+                else if (message.error !== undefined)
+                    ToasterError(message.error);
+                else if (message.user !== undefined)
+                    ToasterError(message.user);
                 else
-                    refresh();
+                    props.refresh();
             })
     }
 
     const deleteMatchButtonHandler = () => {
+        if (props.value.id < 0){
+            ToasterError("Id needs to be a positive integer");
+            return;
+        }
+
         const requestOptions = {
             method: 'DELETE'
         };
@@ -372,13 +490,7 @@ export default function CustomForm(props) {
         const URL = URL_BASE + String(matchValue.id) + "/clubs/"
 
         fetch(URL, requestOptions)
-            .then(message => message.json())
-            .then((message) => {
-                if (message.error !== undefined)
-                    ToasterError(message.error[0]);
-                else
-                    refresh();
-            })
+            .then(() => {props.refresh();});
     }
 
     return (
@@ -465,6 +577,7 @@ export default function CustomForm(props) {
                         setPageNumber = {setPageNumber}
                         paginationOptions = {paginationValue}
                         paginationHandler = {setPaginationValue}
+                        userClickHandler = {props.userClickHandler}
                     ></CustomTable>
                 </Container>
             }
